динамическое подключение элементов формы

+++ core/admin/views/add.php
@@ -6,11 +6,13 @@
                     <input type="submit" class="vg-text vg-firm-color1 vg-firm-background-color4 vg-input vg-button" value="Сохранить">
                 </div>
 
-                <div class="vg-element vg-padding-in-px">
-                    <a href="/admin/delete/goods/53" class="vg-text vg-firm-color1 vg-firm-background-color4 vg-input vg-button vg-center vg_delete">
-                        <span>Удалить</span>
-                    </a>
-                </div>
+                <?php if(!$this->noDelete && $this->data): ?>
+                    <div class="vg-element vg-padding-in-px">
+                        <a href="/admin/delete/goods/53" class="vg-text vg-firm-color1 vg-firm-background-color4 vg-input vg-button vg-center vg_delete">
+                            <span>Удалить</span>
+                        </a>
+                    </div>
+                <?php endif; ?>
             </div>
         </div>
     </div>
@@ -19,24 +21,38 @@
         <input id="tableId" type="hidden" name="<?= $this->columns['id_row']?>" value="<?= $this->data[$this->columns['id_row']] ?>">
     <?php endif; ?>
 
-
     <input type="hidden" name="table" value="<?= $this->table ?>">
 
-    <div class="vg-wrap vg-element vg-rows">
-        <div class="vg-full vg-firm-background-color4 vg-box-shadow">
+    <?php
+        foreach($this->blocks as $class =>$block){
 
-        </div>
-    </div>
+            if(is_int($class)) $class = 'vg-content';
 
-    <div class="vg-wrap vg-element vg-img">
-        <div class="vg-full vg-firm-background-color4 vg-box-shadow">
+           echo '<div class="vg-wrap vg-element '. $class .'">';
 
-        </div>
-    </div>
+           echo '<div class="vg-full vg-firm-background-color4 vg-box-shadow">';
 
-    <div class="vg-wrap vg-element vg-content">
+           if($block){
 
-    </div>
+               foreach($block as $row){
+
+                    foreach ($this->templateArr as $template => $items){
+
+                        if(in_array($row,$items)){
+                            if(!@include $_SERVER['DOCUMENT_ROOT'] . $this->formTemplates . $template . '.php'){
+                                throw new \core\base\exceptions\RouteException('Не найден шаблон '.
+                                    $_SERVER['DOCUMENT_ROOT'] . $this->formTemplates . $template . '.php');
+                            }
+                            break;
+                        }
+                    }
+               }
+           }
+           echo '</div>';
+           echo '</div>';
+        }
+    ?>
+
 
     <div class="vg-wrap vg-element vg-full">
         <div class="vg-wrap vg-element vg-full vg-firm-background-color4 vg-box-shadow">
@@ -44,12 +60,15 @@
                 <div class="vg-element vg-padding-in-px">
                     <input type="submit" class="vg-text vg-firm-color1 vg-firm-background-color4 vg-input vg-button" value="Сохранить">
                 </div>
-                <div class="vg-element vg-padding-in-px">
-                    <a href="/admin/shop/delete/table/shop_products/id_row/id/id/92" class="vg-text vg-firm-color1 vg-firm-background-color4 vg-input vg-button vg-center vg_delete">
-                        <span>Удалить</span>
-                    </a>
-                </div>
+                <?php if(!$this->noDelete && $this->data): ?>
+                    <div class="vg-element vg-padding-in-px">
+                        <a href="/admin/delete/goods/53" class="vg-text vg-firm-color1 vg-firm-background-color4 vg-input vg-button vg-center vg_delete">
+                            <span>Удалить</span>
+                        </a>
+                    </div>
+                <?php endif; ?>
             </div>
         </div>
     </div>
+
 </form>
